Minor BugFixes

+++ MarbleGame/Game/CMMarbleScoreModeScore.m
@@ -93,7 +93,7 @@
     [self.gameDelegate triggerEffect:kCMMarbleEffect_Explode atPosition:hitPosition];
     [self.gameDelegate triggerEffect:kCMMarbleEffect_ComboHit atPosition:hitPosition];
 		comboMultiplier = MARBLE_COMBO_MULTIPLYER * (self.comboHits -1);
-    NSLog(@"ComboHits: %ld" ,self.comboHits);
+    NSLog(@"ComboHits: %ld" ,(long)self.comboHits);
 //		self.comboHits -= [removedMarbles count];
 	}
